:art: updated game service and controller and updated routes
