+++ apps/web/src/routes/shifumi.$gameId.tsx
@@ -1,0 +1,9 @@
+import { createFileRoute } from '@tanstack/react-router'
+
+export const Route = createFileRoute('/shifumi/$gameId')({
+  component: RouteComponent,
+})
+
+function RouteComponent() {
+  return <div>Hello "/shifumi/$gameId"!</div>
+}
